tektronix: Move tektronix agent into package
