+++ docker-compose.yml
@@ -55,15 +55,6 @@ services:
       - "socs"
 
   # --------------------------------------------------------------------------
-  # Tektronix3021c
-  # --------------------------------------------------------------------------
-  ocs-tektronix3021c-agent:
-    image: "ocs-tektronix3021c-agent"
-    build: ./agents/tektronix3021c/
-    depends_on:
-      - "socs"
-
-  # --------------------------------------------------------------------------
   # Pfeiffer Turbo Controller
   # --------------------------------------------------------------------------
   ocs-pfeiffer-tc400-agent:
