+++ tests/agents/test_tektronix_agent.py
@@ -1,3 +1,1 @@
-import sys
-sys.path.insert(0, '../agents/tektronix3021c/')
-from tektronix_agent import TektronixAWGAgent  # noqa: F401
+from socs.agents.tektronix3021c.agent import TektronixAWGAgent  # noqa: F401
